separate flow and similarity cli
