+++ projects/train/configs/flow/sg.yaml
@@ -8,7 +8,7 @@ trainer:
   logger:
     - class_path: lightning.pytorch.loggers.WandbLogger
       init_args:
-        save_dir: "/home/ethan.marx/amplfi/runs/first-lightning-run/train/"
+        save_dir: ${oc.env:AMPLFI_OUTDIR}/train
   max_epochs: 200
   check_val_every_n_epoch: 1
   log_every_n_steps: 20
@@ -16,7 +16,7 @@ trainer:
 model:
   class_path: train.models.flow.FlowModel
   init_args:
-    outdir: "/home/ethan.marx/amplfi/runs/first-lightning-run/"
+    outdir: ${oc.env:AMPLFI_OUTDIR}
     arch:
       class_path: amplfi.architectures.flows.MaskedAutoRegressiveFlow
       init_args:
@@ -43,7 +43,7 @@ model:
 data:
   class_path: train.data.datasets.FlowDataset
   init_args:
-    data_dir: "/home/ethan.marx/projects/PE/projects/train/datadir_tmp/"
+    data_dir: ${oc.env:AMPLFI_DATADIR}
     inference_params: ["frequency", "quality", "hrss", "phase",  "eccentricity", "dec", "psi", "phi"]
     highpass: 25
     sample_rate: 2048
