+++ projects/train/train/cli/similarity.py
@@ -1,0 +1,34 @@
+from train.cli.base import AmplfiBaseCLI
+from train.data.datasets.base import AmplfiDataset
+from train.models.base import AmplfiModel
+
+
+class AmplfiSimilarityCli(AmplfiBaseCLI):
+    def add_arguments_to_parser(self, parser):
+        super().add_arguments_to_parser(parser)
+
+        parser.link_arguments(
+            "data.init_args.ifos",
+            "model.init_args.arch.init_args.num_ifos",
+            compute_fn=lambda x: len(x),
+            apply_on="parse",
+        )
+        return parser
+
+
+def main(args=None):
+    cli = AmplfiSimilarityCli(
+        AmplfiModel,
+        AmplfiDataset,
+        subclass_mode_model=True,
+        subclass_mode_data=True,
+        save_config_kwargs={"overwrite": True},
+        seed_everything_default=101588,
+        args=args,
+        parser_kwargs={"parser_mode": "omegaconf"},
+    )
+    return cli
+
+
+if __name__ == "__main__":
+    main()
